[User list] Update user type filter
Updated usertype filter to include the "Global" option.
Corrected a bug that when selecting the userType "Inter-Agency"
as a filter. The results would always be zero as it searched
for an incorrect usergroup name. This is corrected by adding
a specific value to the secondary filter options that will be
used over the name if it is available.
Minor, updated the versions of gulp-ng-annotate to work properly
with io.js.

References jqueipo/DATIM#427